Move runner to core module and create system threads
